# gen/docs/example.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "example.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "ExampleService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/service/name": {
      "get": {
        "operationId": "ExampleService_ServiceName",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ServiceNameResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "tags": [
          "ExampleService"
        ]
      }
    },
    "/v1/service/status": {
      "post": {
        "operationId": "ExampleService_Status",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1StatusResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1StatusRequest"
            }
          }
        ],
        "tags": [
          "ExampleService"
        ]
      }
    }
  },
  "definitions": {
    "examplev1Status": {
      "type": "string",
      "enum": [
        "UNKNOWN",
        "REGISTERED",
        "NOT_REGISTERED"
      ],
      "default": "UNKNOWN"
    },
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "v1ServiceNameResponse": {
      "type": "object",
      "properties": {
        "Name": {
          "type": "string"
        }
      }
    },
    "v1StatusRequest": {
      "type": "object",
      "properties": {
        "ServiceName": {
          "type": "string"
        }
      }
    },
    "v1StatusResponse": {
      "type": "object",
      "properties": {
        "Status": {
          "$ref": "#/definitions/examplev1Status"
        }
      }
    }
  }
}